rename: template 폴더 내에 있는 파일명 수정
