+++ src/components/templates/club/About.tsx
@@ -1,17 +1,17 @@
 import Heading from "~/components/atoms/club/Heading";
-import { AboutPageLayout } from "./AboutPageStyles";
+import { AboutLayout } from "./AboutStyles";
 import helmet from "~/assets/images/helmet.svg";
 import AboutTeam from "~/components/atoms/club/about/AboutTeam";
 import History from "~/components/atoms/club/about/History";
 import HISTORY_DATA from "../../../constants/historyData";
 
 const About = () => (
-  <AboutPageLayout>
+  <AboutLayout>
     <Heading src={helmet} alt="helmet" text="개요" />
     <AboutTeam />
     <Heading src={helmet} alt="helmet" text="연혁" />
     <History historyData={HISTORY_DATA} />
-  </AboutPageLayout>
+  </AboutLayout>
 );
 
 export default About;
