+++ src/components/templates/club/AboutStyles.ts
@@ -1,0 +1,12 @@
+import tw from "twin.macro";
+import styled from "styled-components";
+
+export const AboutLayout = styled.div`
+  ${tw`
+    font-bold
+    h-full
+    w-full
+    px-32
+    py-20
+  `}
+`;
