+++ src/components/templates/club/CoachStyles.ts
@@ -1,12 +1,18 @@
 import tw from "twin.macro";
 import styled from "styled-components";
 
-export const AboutPageLayout = styled.div`
+export const CoachLayout = styled.div`
   ${tw`
     font-bold
     h-full
     w-full
     px-32
     py-20
+  `}
+`;
+
+export const CoachListWrap = styled.div`
+  ${tw`
+    py-10
   `}
 `;
