--- src/components/templates/club/OrganizationPageStyles.ts
@@ -1,17 +1,0 @@
-import tw from "twin.macro";
-import styled from "styled-components";
-
-export const OrganizationPageLayout = styled.div`
-  ${tw`
-    h-full
-    w-full
-    pt-20
-    px-32
-  `}
-`;
-
-export const OrganizationChartWrap = styled.div`
-  ${tw`
-    my-20
-  `}
-`;
