+++ src/components/templates/club/Sponsor.tsx
@@ -1,17 +1,17 @@
-import { SponsorPageLayout, SponsorList } from "./SponsorPageStyles";
+import { SponsorLayout, SponsorList } from "./SponsorStyles";
 import Company from "~/components/atoms/club/sponsor/Company";
 import LasanLogo from "~/assets/images/lasangroup.svg";
 import IlogisticLogo from "~/assets/images/ilogistic.svg";
 
 const Sponsor = () => {
   return (
-    <SponsorPageLayout>
+    <SponsorLayout>
       <SponsorList>
         <Company SponsorNameKo="라산 그룹" SponsorNameEn="LASAN GROUP" src={LasanLogo} />
         <Company SponsorNameKo="아이로지스틱" SponsorNameEn="ILOGISTIC" src={IlogisticLogo} />
         <Company SponsorNameKo="테스트" SponsorNameEn="테스트" src={IlogisticLogo} />
       </SponsorList>
-    </SponsorPageLayout>
+    </SponsorLayout>
   );
 };
