+++ src/pages/club/AboutPage.tsx
@@ -1,4 +1,4 @@
-import About from "~/components/templates/club/AboutPage";
+import About from "~/components/templates/club/About";
 
 const AboutPage = () => {
   return <About />;
